backend/fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
PyMySQL==1.1.0
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-decouple==3.8
emails==0.6
Jinja2==3.1.2
aioredis==2.0.1
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pyotp==2.9.0
qrcode[pil]==7.4.2
python-social-auth==0.3.6
social-auth-app-fastapi==0.1.0
websockets==12.0
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
pillow==10.1.0
requests==2.31.0
aiofiles==23.2.1
tenacity==8.2.3
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1